# core-common-security-vault-sdk/src/main/resources/api-spec/openapi.yml
openapi: 3.0.1
info:
  title: core-common-security-vault
  description: Common Platform Security Vault - Secure Credential Storage
  contact:
    name: Firefly Software Foundation
    email: dev@getfirefly.io
  version: 1.0.0
servers:
  - url: http://core.getfirefly.io/vault
    description: Development Environment
  - url: /
    description: Local Development Environment
...
paths:
...
  /api/v1/credentials/{id}:
    get:
      tags:
        - Credentials
      summary: Get a credential by ID
      description: Returns a credential based on the ID
      operationId: getCredentialById
      parameters:
        - name: id
          in: path
          description: ID of the credential to retrieve
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CredentialDTO'
        '404':
          description: Credential not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CredentialDTO'
    put:
      tags:
        - Credentials
      summary: Update an existing credential
      description: Updates an existing credential and returns it
      operationId: updateCredential
      parameters:
        - name: id
          in: path
          description: ID of the credential to update
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CredentialDTO'
        required: true
      responses:
        '200':
          description: Credential updated
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CredentialDTO'
        '400':
          description: Invalid input
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CredentialDTO'
        '404':
          description: Credential not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CredentialDTO'
    delete:
      tags:
        - Credentials
      summary: Delete a credential
      description: Deletes a credential
      operationId: deleteCredential
      parameters:
        - name: id
          in: path
          description: ID of the credential to delete
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '204':
          description: Credential deleted
        '404':
          description: Credential not found
  /api/v1/credential-types/{id}:
    get:
      tags:
        - Credential Types
      summary: Get a credential type by ID
      operationId: getCredentialTypeById
      parameters:
        - name: id
          in: path
          description: ID of the credential type to retrieve
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Credential type found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CredentialTypeDTO'
        '404':
          description: Credential type not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CredentialTypeDTO'
    put:
      tags:
        - Credential Types
      summary: Update an existing credential type
      operationId: updateCredentialType
      parameters:
        - name: id
          in: path
          description: ID of the credential type to update
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CredentialTypeDTO'
        required: true
      responses:
        '200':
          description: Credential type updated
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CredentialTypeDTO'
        '404':
          description: Credential type not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CredentialTypeDTO'
    delete:
      tags:
        - Credential Types
      summary: Delete a credential type
      operationId: deleteCredentialType
      parameters:
        - name: id
          in: path
          description: ID of the credential type to delete
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '204':
          description: Credential type deleted
        '404':
          description: Credential type not found
...
  /api/v1/credentials/filter:
    post:
      tags:
        - Credentials
      summary: Filter credentials
      description: Returns a filtered list of credentials based on criteria
      operationId: filterCredentials
      parameters:
        - name: pagination.pageNumber
          in: query
          description: Page number (0-based)
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: pagination.pageSize
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            format: int32
            default: 10
        - name: pagination.sortBy
          in: query
          description: Field to sort by
          required: false
          schema:
            type: string
        - name: pagination.sortDirection
          in: query
          description: Sort direction (ASC or DESC)
          required: false
          schema:
            type: string
            default: DESC
        - name: options.caseInsensitiveStrings
          in: query
          description: Enable case-insensitive string filtering
          required: false
          schema:
            type: boolean
            default: false
        - name: options.includeInheritedFields
          in: query
          description: Include fields from parent classes
          required: false
          schema:
            type: boolean
            default: false
        - name: filters.code
          in: query
          description: Exact filter for code
          required: false
          schema:
            type: string
        - name: filters.name
          in: query
          description: Exact filter for name
          required: false
          schema:
            type: string
        - name: filters.description
          in: query
          description: Exact filter for description
          required: false
          schema:
            type: string
        - name: filters.encryptedValue
          in: query
          description: Exact filter for encrypted value
          required: false
          schema:
            type: string
        - name: filters.encryptionAlgorithm
          in: query
          description: Exact filter for encryption algorithm
          required: false
          schema:
            type: string
        - name: filters.encryptionIv
          in: query
          description: Exact filter for encryption iv
          required: false
          schema:
            type: string
        - name: filters.credentialOwner
          in: query
          description: Exact filter for credential owner
          required: false
          schema:
            type: string
        - name: filters.credentialContactEmail
          in: query
          description: Exact filter for credential contact email
          required: false
          schema:
            type: string
        - name: filters.expiresAt
          in: query
          description: Exact filter for expires at
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[expiresAt].from
          in: query
          description: Filter expires at from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[expiresAt].to
          in: query
          description: Filter expires at to value
          required: false
          schema:
            type: string
            format: date-time
        - name: filters.rotateBeforeDays
          in: query
          description: Exact filter for rotate before days
          required: false
          schema:
            type: integer
            format: int32
        - name: rangeFilters.ranges[rotateBeforeDays].from
          in: query
          description: Filter rotate before days from value
          required: false
          schema:
            type: integer
            format: int32
        - name: rangeFilters.ranges[rotateBeforeDays].to
          in: query
          description: Filter rotate before days to value
          required: false
          schema:
            type: integer
            format: int32
        - name: filters.lastRotatedAt
          in: query
          description: Exact filter for last rotated at
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[lastRotatedAt].from
          in: query
          description: Filter last rotated at from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[lastRotatedAt].to
          in: query
          description: Filter last rotated at to value
          required: false
          schema:
            type: string
            format: date-time
        - name: filters.rotationEnabled
          in: query
          description: Exact filter for rotation enabled
          required: false
          schema:
            type: boolean
        - name: filters.autoRotationDays
          in: query
          description: Exact filter for auto rotation days
          required: false
          schema:
            type: integer
            format: int32
        - name: rangeFilters.ranges[autoRotationDays].from
          in: query
          description: Filter auto rotation days from value
          required: false
          schema:
            type: integer
            format: int32
        - name: rangeFilters.ranges[autoRotationDays].to
          in: query
          description: Filter auto rotation days to value
          required: false
          schema:
            type: integer
            format: int32
        - name: filters.lastUsedAt
          in: query
          description: Exact filter for last used at
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[lastUsedAt].from
          in: query
          description: Filter last used at from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[lastUsedAt].to
          in: query
          description: Filter last used at to value
          required: false
          schema:
            type: string
            format: date-time
        - name: filters.usageCount
          in: query
          description: Exact filter for usage count
          required: false
          schema:
            type: integer
            format: int32
        - name: rangeFilters.ranges[usageCount].from
          in: query
          description: Filter usage count from value
          required: false
          schema:
            type: integer
            format: int32
        - name: rangeFilters.ranges[usageCount].to
          in: query
          description: Filter usage count to value
          required: false
          schema:
            type: integer
            format: int32
        - name: filters.lastAccessedBy
          in: query
          description: Exact filter for last accessed by
          required: false
          schema:
            type: string
        - name: filters.accessScope
          in: query
          description: Exact filter for access scope
          required: false
          schema:
            type: string
        - name: filters.allowedServices
          in: query
          description: Exact filter for allowed services
          required: false
          schema:
            type: string
        - name: filters.allowedIps
          in: query
          description: Exact filter for allowed ips
          required: false
          schema:
            type: string
        - name: filters.allowedEnvironments
          in: query
          description: Exact filter for allowed environments
          required: false
          schema:
            type: string
        - name: filters.isSensitive
          in: query
          description: Exact filter for is sensitive
          required: false
          schema:
            type: boolean
        - name: filters.requireApprovalForAccess
          in: query
          description: Exact filter for require approval for access
          required: false
          schema:
            type: boolean
        - name: filters.auditAllAccess
          in: query
          description: Exact filter for audit all access
          required: false
          schema:
            type: boolean
        - name: filters.maskInLogs
          in: query
          description: Exact filter for mask in logs
          required: false
          schema:
            type: boolean
        - name: filters.backupEnabled
          in: query
          description: Exact filter for backup enabled
          required: false
          schema:
            type: boolean
        - name: filters.backupLocation
          in: query
          description: Exact filter for backup location
          required: false
          schema:
            type: string
        - name: filters.tags
          in: query
          description: Exact filter for tags
          required: false
          schema:
            type: string
        - name: filters.metadata
          in: query
          description: Exact filter for metadata
          required: false
          schema:
            type: string
        - name: filters.active
          in: query
          description: Exact filter for active
          required: false
          schema:
            type: boolean
        - name: filters.version
          in: query
          description: Exact filter for version
          required: false
          schema:
            type: integer
            format: int32
        - name: rangeFilters.ranges[version].from
          in: query
          description: Filter version from value
          required: false
          schema:
            type: integer
            format: int32
        - name: rangeFilters.ranges[version].to
          in: query
          description: Filter version to value
          required: false
          schema:
            type: integer
            format: int32
        - name: filters.createdAt
          in: query
          description: Exact filter for created at
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[createdAt].from
          in: query
          description: Filter created at from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[createdAt].to
          in: query
          description: Filter created at to value
          required: false
          schema:
            type: string
            format: date-time
        - name: filters.updatedAt
          in: query
          description: Exact filter for updated at
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[updatedAt].from
          in: query
          description: Filter updated at from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[updatedAt].to
          in: query
          description: Filter updated at to value
          required: false
          schema:
            type: string
            format: date-time
        - name: filters.createdBy
          in: query
          description: Exact filter for created by
          required: false
          schema:
            type: string
        - name: filters.updatedBy
          in: query
          description: Exact filter for updated by
          required: false
          schema:
            type: string
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PaginationResponseCredentialDTO'
...
  /api/v1/credential-types:
    post:
      tags:
        - Credential Types
      summary: Create a new credential type
      operationId: createCredentialType
      parameters:
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CredentialTypeDTO'
        required: true
      responses:
        '201':
          description: Credential type created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CredentialTypeDTO'
        '400':
          description: Invalid input
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CredentialTypeDTO'
  /api/v1/credential-types/filter:
    post:
      tags:
        - Credential Types
      summary: Filter credential types
      operationId: filterCredentialTypes
      parameters:
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FilterRequestCredentialTypeDTO'
        required: true
      responses:
        '200':
          description: Credential types filtered successfully
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PaginationResponseCredentialTypeDTO'
...
  /api/v1/credential-versions/credential/{credentialId}:
    get:
      tags:
        - Credential Versions
      summary: Get all versions for a credential
      operationId: getVersionsByCredentialId
      parameters:
        - name: credentialId
          in: path
          description: Credential ID
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CredentialVersionDTO'
...
components:
  schemas:
...
    CredentialDTO:
      required:
        - code
        - credentialStatusId
        - credentialTypeId
        - encryptedValue
        - environmentTypeId
        - name
      type: object
      properties:
        id:
          type: string
          description: Credential ID
          format: uuid
          readOnly: true
        code:
          maxLength: 100
          minLength: 2
          type: string
          description: Unique code for the credential
          example: STRIPE_API_KEY_PROD
        name:
          maxLength: 255
          minLength: 2
          type: string
          description: Name of the credential
          example: Stripe Production API Key
        description:
          type: string
          description: Description of the credential
          example: API key for Stripe payment gateway in production
        credentialTypeId:
          type: string
          description: Credential type ID
          format: uuid
        credentialStatusId:
          type: string
          description: Credential status ID
          format: uuid
        environmentTypeId:
          type: string
          description: Environment type ID
          format: uuid
        providerId:
          type: string
          description: Provider ID
          format: uuid
        integrationId:
          type: string
          description: Integration ID
          format: uuid
        serviceId:
          type: string
          description: Service ID
          format: uuid
        tenantId:
          type: string
          description: Tenant ID
          format: uuid
        encryptedValue:
          type: string
          description: Encrypted credential value
          writeOnly: true
        encryptionAlgorithm:
          type: string
          description: Encryption algorithm
          example: AES-256-GCM
        encryptionKeyId:
          type: string
          description: Encryption key ID
        encryptionIv:
          type: string
          description: Encryption IV
        credentialOwner:
          type: string
          description: Credential owner
        credentialContactEmail:
          type: string
          description: Credential contact email
        expiresAt:
          type: string
          description: Expiration timestamp
          format: date-time
        rotateBeforeDays:
          type: integer
          description: Rotate before days
          format: int32
        lastRotatedAt:
          type: string
          description: Last rotated timestamp
          format: date-time
        rotationEnabled:
          type: boolean
          description: Rotation enabled
        autoRotationDays:
          type: integer
          description: Auto rotation days
          format: int32
        lastUsedAt:
          type: string
          description: Last used timestamp
          format: date-time
          readOnly: true
        usageCount:
          type: integer
          description: Usage count
          format: int64
          readOnly: true
        lastAccessedBy:
          type: string
          description: Last accessed by
          readOnly: true
        accessScope:
          type: string
          description: Access scope
          example: INTERNAL
        allowedServices:
          type: string
          description: Allowed services
        allowedIps:
          type: string
          description: Allowed IPs
        allowedEnvironments:
          type: string
          description: Allowed environments
        isSensitive:
          type: boolean
          description: Is sensitive
        requireApprovalForAccess:
          type: boolean
          description: Require approval for access
        auditAllAccess:
          type: boolean
          description: Audit all access
        maskInLogs:
          type: boolean
          description: Mask in logs
        backupEnabled:
          type: boolean
          description: Backup enabled
        backupLocation:
          type: string
          description: Backup location
        tags:
          type: string
          description: Tags
        metadata:
          type: string
          description: Metadata
        active:
          type: boolean
          description: Whether the credential is active
          default: true
        version:
          type: integer
          description: Version for optimistic locking
          format: int64
          readOnly: true
        createdAt:
          type: string
          description: Creation timestamp
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          description: Last update timestamp
          format: date-time
          readOnly: true
        createdBy:
          type: string
          description: Created by user ID
          format: uuid
          readOnly: true
        updatedBy:
          type: string
          description: Updated by user ID
          format: uuid
          readOnly: true
      description: Credential information
    CredentialTypeDTO:
      required:
        - code
        - name
      type: object
      properties:
        id:
          type: string
          description: Credential type ID
          format: uuid
          readOnly: true
        code:
          maxLength: 50
          minLength: 2
          type: string
          description: Unique code for the credential type
          example: API_KEY
        name:
          maxLength: 100
          minLength: 2
          type: string
          description: Name of the credential type
          example: API Key
        description:
          type: string
          description: Description of the credential type
          example: Simple API key authentication
        category:
          type: string
          description: Category of the credential type
          example: AUTHENTICATION
        active:
          type: boolean
          description: Whether the credential type is active
          default: true
        version:
          type: integer
          description: Version for optimistic locking
          format: int64
          readOnly: true
        createdAt:
          type: string
          description: Creation timestamp
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          description: Last update timestamp
          format: date-time
          readOnly: true
      description: Credential type to create
...
    RangeFilter:
      type: object
      properties:
        ranges:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/RangeObject'
          description: Map of field names to their range criteria
      description: Represents range filters for numeric fields
...
    PaginationResponseCredentialDTO:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            $ref: '#/components/schemas/CredentialDTO'
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: The current page number, typically zero-based.
          format: int32
      description: Represents a paginated response containing a list of items and pagination metadata.
...
    FilterRequestCredentialTypeDTO:
      required:
        - pagination
      type: object
      properties:
        filters:
          $ref: '#/components/schemas/CredentialTypeDTO'
        rangeFilters:
          $ref: '#/components/schemas/RangeFilter'
        pagination:
          $ref: '#/components/schemas/PaginationRequest'
        options:
          $ref: '#/components/schemas/FilterOptions'
      description: Filter criteria
    PaginationResponseCredentialTypeDTO:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            $ref: '#/components/schemas/CredentialTypeDTO'
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: The current page number, typically zero-based.
          format: int32
      description: Represents a paginated response containing a list of items and pagination metadata.
...